Hacky make dissertation-final-library.pdf.
The goal was to have the date "22. September 2018" twice on page 3. This was not straightforward done in LyX, because although zweihorn has LyX installed, it does not have LaTeX installed. And Knoppix 8.1 DVD, which has LaTeX installed, does not have LyX installed.
I first generated the 3rd page using pdflatex, and used the program pdfchain to merge 3 pdf files: pages1-2 of dissertation-final-7beaae.pdf, the new 3rd page, and pages4-163 of dissertation-final-7beaae.pdf. But the resulting pdf file did not have links anymore (the Table of Contents, Bibliography links, and Index did not have red or green boxes around them in evince). Therefore the following was carried out:

1. Edit using LyX, title.lyx to show the date "22. September 2018" twice.
2. Export (to pdflatex) dissertation.lyx on zweihorn to tex-files (which currently doesn't have dia installed). Copy onto the USB-Stick and start Knoppix 8.1 DVD.
3. On Knoppix 8.1 DVD, start Dia, edit 3 .dia files, which did not have a corresponding pdf file (using File/Page Setup/, setting all Margins to 0cm, setting Scaling to Fit to: 1 by 1, clicking OK), and export the pdf files using Dia's Cairo PDF.
4. Edit the \begin{figure} sections of the tex files: remove eps files, instead include the pdf files, and remove the "bb = 0 0 200 100, draft, type=eps" parts of all \includegraphics.
5. Rename zusammenfassung.bib to dissertation.bib, which was necessary for pdflatex to find the bibliography created my bibtex.
6. Then write a Makefile that runs pdflatex, bibtex, and makeindex in the correct order to produce dissertation.pdf.
7. Verify that the pages, where modified .dia (and exported .pdf) files are, end with the same word as the printed copies. The only visible difference is that the 3 .dia graphics are not centered, but on the left side. I tried to center the shapes into the middle of the page, but Dia always adjusted the left border. I think this is due to Dia being version "dia 0.97+git". Anyways, I think the resulting dissertation-final-library.pdf is good for submission to the library tomorrow (today actually).